+++ changelog.md
@@ -1,6 +1,7 @@
 ## 17.04.2023, Mon
 
 - 15:00 - Frontend: implemented "Manage folders" page UI.
+- 19:40 - Backend: created endpoints to create, update and delete user folders.
 
 ## 16.04.2023, Sun